Make LoudnessAnalyzer cancellable and widen the frame counters
analyze() was a synchronous C decode loop with no cancellation hook, reached
through an async init that awaits nothing — so cancelling Detect Loudness on a
long file never returned and the progress sheet could only be escaped by force
quitting. It now takes an isCancelled predicate, defaulting to the calling
task, and throws CancellationError.

The predicate is polled once per converter iteration rather than inside the
callback. Storing it on CallbackContext makes the struct non-trivial, and
handing that to AudioConverterFillComplexBuffer as untyped memory warns about
forming a raw pointer to a type holding an object reference. One iteration is
192,000 source frames, measured at roughly 66 ms on the reference file.

framesProduced accumulated oversampled frames into a UInt32 and was never read,
trapping at about 89.5% of a 6.9-hour file. Deleted. fileFramesRead is widened
to Int64 to match fileLengthInFrames and targetFrames.

NormalizeAnalyzer now rethrows CancellationError instead of collapsing it into
a unity-gain result, which the hook above would otherwise have made reachable.

The overflow fix has no committed test: reproducing it needs a 4.8 GB fixture,
which cannot be checked in. Verified once from a throwaway suite against
scratchpad/rf64-testing/big-rf64.wav — 1,200,000,000 frames analyzed to
completion in 410.7 s where it previously trapped. That path is uncovered.

+++ Sources/SPFKLoudness/Internal/AudioConverterCallback.swift
@@ -55,7 +55,7 @@ let audioConverterCallback: AudioConverterComplexInputDataProc = {
         if rereadErr != noErr { return rereadErr }
     }
 
-    context.pointee.fileFramesRead += framesInFileOutBuffer
+    context.pointee.fileFramesRead += Int64(framesInFileOutBuffer)
 
     var framesInBuffer = framesInFileOutBuffer
     var pushedFrames: UInt32 = 0
